**BREAKING JAX-RS CHANGE** Fixed bugs in parameter conversions for JAX-RS methods to match the spec. This may affect cases where empty values were being sent via parameters (e.g. `param` will be considered an empty string rather than null in a request such as `/path?param=`) and custom param type converters for Collections now behave differently: if there is a type converter for type `A` then the same converter will be used for parameters of type `A` or `List<A>` etc. It is not currently possible to specify a custom List, Set, SortedSet, or Collection provider (but other collection types can be specified).
Finally, parameters that had commas in them will no longer be split into a list, so `?param=one,two` will now be considered to have a single value of `one,two`.

Although breaking changes are generally introduced by deprecating old members and keeping backward compatibility, in this case it was not possible due to the fact that these are bug fixes for parameter handling as described in the JAX-RS spec.

If the non-standard behaviour is relied on (e.g. you have clients sending empty values to indicate `null` in query strings) you can add post-match filters that delete any empty string parameters from the headers or query strings. To treat commas in parameter values as lists a similar post-match filter could split any relevant values on the comma and generate extra parameters.